+++ src/eventsub/mod.rs
@@ -164,6 +164,10 @@ pub enum Payload {
     ChannelRaidV1(NotificationPayload<channel::ChannelRaidV1>),
     /// Channel Subscription End V1 Event
     ChannelSubscriptionEndV1(NotificationPayload<channel::ChannelSubscriptionEndV1>),
+    /// Channel Subscription Gift V1 Event
+    ChannelSubscriptionGiftV1(NotificationPayload<channel::ChannelSubscriptionGiftV1>),
+    /// Channel Subscription Message V1 Event
+    ChannelSubscriptionMessageV1(NotificationPayload<channel::ChannelSubscriptionMessageV1>),
 }
 
 impl Payload {
@@ -359,6 +363,8 @@ impl<'de> Deserialize<'de> for Payload {
                 channel::ChannelPredictionEndV1;
                 channel::ChannelRaidV1;
                 channel::ChannelSubscriptionEndV1;
+                channel::ChannelSubscriptionGiftV1;
+                channel::ChannelSubscriptionMessageV1;
                 channel::ChannelHypeTrainBeginV1;
                 channel::ChannelHypeTrainProgressV1;
                 channel::ChannelHypeTrainEndV1;
@@ -397,6 +403,8 @@ impl<'de> Deserialize<'de> for Payload {
                 channel::ChannelPredictionEndV1;
                 channel::ChannelRaidV1;
                 channel::ChannelSubscriptionEndV1;
+                channel::ChannelSubscriptionGiftV1;
+                channel::ChannelSubscriptionMessageV1;
                 channel::ChannelHypeTrainBeginV1;
                 channel::ChannelHypeTrainProgressV1;
                 channel::ChannelHypeTrainEndV1;
@@ -548,6 +556,12 @@ pub enum EventType {
     /// `channel.subscription.end`: a subscription to the specified channel expires.
     #[serde(rename = "channel.subscription.end")]
     ChannelSubscriptionEnd,
+    /// `channel.subscription.gift`: a user gives one or more gifted subscriptions in a channel.
+    #[serde(rename = "channel.subscription.gift")]
+    ChannelSubscriptionGift,
+    /// `channel.subscription.gift`: a user sends a resubscription chat message in a specific channel
+    #[serde(rename = "channel.subscription.message")]
+    ChannelSubscriptionMessage,
     /// `channel.hype_train.begin`: a hype train begins on the specified channel.
     #[serde(rename = "channel.hype_train.begin")]
     ChannelHypeTrainBegin,